Testing

+++ test.sh
@@ -3,9 +3,9 @@
 server_port="80"
 open_screenshots=0
 run_test=all
-margarita_pid=0
 file=0
 integ_ext="coffee"
+force=0
 
 usage() {
   cat 1>&2 << EOF
@@ -19,6 +19,8 @@ options:
   -m, --start-margarita: Starts the margarita server. See --setup-margarita
   -f, --file name   : Runs this test file only (use with -t, works only with integration tests).
                       Loaded file is "integration/test-{name}.coffee"
+  --force           : Run test even if sanity checks failed
+  --seed [prompt]   : Prepares the database with fresh test data (prompts for password if prompt parameter present)
   --precompile      : Precompiles coffee integration test files (try this if first test hangs)
   --setup-margarita : Downloads the margarita server, installs the profile and dependencies and exits
 EOF
@@ -41,24 +43,33 @@ do_api_test() {
 
 setup_margarita() {
   git clone https://github.com/CourseAdvisor/margarita.git
-  cp tests/profiles.json margarita/
   cd margarita
   npm install
 }
 
+stop_margarita() {
+  margarita_pid=`cat .margarita.pid 2>/dev/null`
+  if [ $? -eq 0 ] && [ ! $margarita_pid -eq "0" ]; then
+    kill $margarita_pid
+  fi
+  rm .margarita.pid
+}
+
 start_margarita() {
+  if [ -e .margarita.pid ]; then
+    stop_margarita
+  fi
   cd margarita
   node bin/www &
   margarita_pid=$!
+  echo "$margarita_pid" > ../.margarita.pid
   echo "margarita process started with pid $margarita_pid"
   sleep 2
   cd ..
 }
 
 cleanup() {
-  if [ ! $margarita_pid -eq "0" ]; then
-    kill $margarita_pid
-  fi
+  stop_margarita
 }
 
 _do_test() {
@@ -74,6 +85,31 @@ _do_test() {
 }
 
 do_test() {
+  # ensures app is not in debug mode
+  os=$(expr substr $(uname -s) 1 10)
+  if [ "$os" = "MINGW32_NT" ] || [ "$os" = "MINGW64_NT" ]; then #window
+    grep "['\"]debug[\"']" app/config/production/app.php | grep -iqw "0\|false"
+    status=$?
+  else
+    echo 'Config::get("app.debug");' | php artisan tinker | grep -q false
+    status=$?
+  fi
+
+  if [ $status -ne 0 ]; then
+    echo "ERR: Tests must run with debug mode off. Please set app.debug to false in your app config file or use --force if you know what you are doing" 1>&2
+    if [ $force -ne 1 ]; then
+      exit 1
+    fi
+  fi
+
+  # ensures profiles are up to date
+  cp tests/profiles.json margarita/
+
+  # ensures tmp dir
+  if [ ! -d tmp ]; then
+    mkdir tmp
+  fi
+
   cat > tmp/tests-config.coffee << __EOF
     module.exports =
       port: $server_port
@@ -98,10 +134,8 @@ x_open() {
   fi
 }
 
-# ensures tmp dir
-if [ ! -d tmp ]; then
-  mkdir tmp
-fi
+# clean compiled files before running tests
+rm -f tests/integration/*.js
 
 while [ $# -ne 0 ]; do
   case "$1" in
@@ -120,6 +154,9 @@ while [ $# -ne 0 ]; do
       shift
       run_test="$1"
     ;;
+    --force)
+      force=1
+    ;;
     --setup-margarita)
       setup_margarita
       exit $?
@@ -134,6 +171,16 @@ while [ $# -ne 0 ]; do
     --file|-f)
       shift
       file="$1"
+    ;;
+    --seed)
+      mysql_args=""
+      if [ "$2" = "prompt" ]; then
+        shift
+        mysql_args="-p"
+      fi
+      echo "Seeding database. Requires to enter sql root password"
+      echo "drop database IF EXISTS courseadvisor; create database courseadvisor;" |
+        cat - ./app/database/database.sql ./app/database/seeds/testing.sql | mysql -uroot $mysql_args
     ;;
     *)
       echo "Unknown parameter: $1"
